Adds endpoint to trigger position simulation
Adds an API endpoint that allows for triggering the position simulation command.
This endpoint is rate-limited and requires a secret key for authentication.
It also prevents overlapping executions of the simulation.

The rate limit is set to 2 requests per minute.

A new setting for the Cron secret is also introduced.