Improves subscription handling and display
Adds logic to prevent users from subscribing to the same plan twice.
This enhances the user experience by disabling the subscribe button
for the current plan and displaying "You already have this".

It also adds a link to the Stripe billing portal for authenticated
users with active subscriptions.

The ads_enabled check has also been removed from the subscribe template.

A function for calculating subscription class has been implemented and
is used to indicate the users current subscription level.